# spec/v0.1.0/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LifecycleStatus": {
        "enum": [
          "active",
          "stale",
          "archived",
          "deleted"
        ],
        "title": "LifecycleStatus",
        "type": "string"
      },
...
      "MemoryCellCreate": {
        "properties": {
          "access_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryAccessPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "content": {
            "$ref": "#/components/schemas/MemoryContent"
          },
          "identity": {
            "$ref": "#/components/schemas/MemoryIdentity"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryProvenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoring": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryScoring"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "$ref": "#/components/schemas/MemoryType"
          }
        },
        "required": [
          "type",
          "content",
          "identity"
        ],
        "title": "MemoryCellCreate",
        "type": "object"
      },
...
      "MemoryType": {
        "enum": [
          "episodic",
          "semantic",
          "procedural"
        ],
        "title": "MemoryType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Agent Memory Protocol reference server implementation",
    "title": "AMP Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/amp/v1/lifecycle/run": {
      "post": {
        "description": "Trigger one lifecycle pass immediately — for cron, ops, or tests.\n\nRuns decay, plus the retention purge when `AMP_PURGE_RETENTION` is on. Gated\non `AMP_ADMIN_TOKEN` because it mutates lifecycle state for every cell in\nstorage, and erases data outright when purging is enabled. Unset token ==\ndisabled (403), not an open endpoint.",
        "operationId": "run_lifecycle_now_amp_v1_lifecycle_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-amp-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Lifecycle Now Amp V1 Lifecycle Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access denied; also returned when the cell does not exist"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Lifecycle Now"
      }
    },
    "/amp/v1/memories": {
      "get": {
        "description": "List memory cells, one page at a time.\n\n`limit` counts cells this caller may read, and `offset` indexes that same\nreadable stream; `has_more` says whether the next page holds anything. The\nresponse used to report `total`, which was the size of the page it had just\nreturned - a number that read like a count of matches and was not one. A real\ncount is not available without scanning the store on every call, so the field\nis gone rather than kept and distrusted.",
        "operationId": "query_memories_amp_v1_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MemoryType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifecycleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-amp-agent-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "x-amp-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query Memories Amp V1 Memories Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "X-AMP-Agent-ID header is required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Memories"
      },
      "post": {
        "operationId": "create_memory_amp_v1_memories_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-amp-agent-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "x-amp-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryCellCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Memory Amp V1 Memories Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "X-AMP-Agent-ID header is required"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The cell exceeds the maximum size advertised at GET /spec"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Memory"
      }
    },
    "/amp/v1/memories/query": {
      "get": {
        "description": "List memory cells, one page at a time.\n\n`limit` counts cells this caller may read, and `offset` indexes that same\nreadable stream; `has_more` says whether the next page holds anything. The\nresponse used to report `total`, which was the size of the page it had just\nreturned - a number that read like a count of matches and was not one. A real\ncount is not available without scanning the store on every call, so the field\nis gone rather than kept and distrusted.",
        "operationId": "query_memories_amp_v1_memories_query_get",
        "parameters": [
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MemoryType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LifecycleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-amp-agent-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "x-amp-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Query Memories Amp V1 Memories Query Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "X-AMP-Agent-ID header is required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query Memories"
      }
    },
    "/amp/v1/memories/search": {
      "post": {
        "operationId": "search_memories_amp_v1_memories_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-amp-agent-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Agent-Id"
            }
          },
          {
            "in": "header",
            "name": "x-amp-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Amp-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Search Memories Amp V1 Memories Search Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "X-AMP-Agent-ID header is required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Memories"
      }
    },
...
  }
}